+++ packages/google/src/google-provider.ts
@@ -1,8 +1,9 @@
 import {
   EmbeddingModelV3,
+  Experimental_VideoModelV3,
+  ImageModelV3,
   LanguageModelV3,
   ProviderV3,
-  ImageModelV3,
 } from '@ai-sdk/provider';
 import {
   FetchFunction,
@@ -24,6 +25,8 @@ import {
   GoogleGenerativeAIImageModelId,
 } from './google-generative-ai-image-settings';
 import { GoogleGenerativeAIImageModel } from './google-generative-ai-image-model';
+import { GoogleGenerativeAIVideoModel } from './google-generative-ai-video-model';
+import { GoogleGenerativeAIVideoModelId } from './google-generative-ai-video-settings';
 
 export interface GoogleGenerativeAIProvider extends ProviderV3 {
   (modelId: GoogleGenerativeAIModelId): LanguageModelV3;
@@ -33,12 +36,17 @@ export interface GoogleGenerativeAIProvider extends ProviderV3 {
   chat(modelId: GoogleGenerativeAIModelId): LanguageModelV3;
 
   /**
-Creates a model for image generation.
- */
+   * Creates a model for image generation.
+   */
   image(
     modelId: GoogleGenerativeAIImageModelId,
     settings?: GoogleGenerativeAIImageSettings,
   ): ImageModelV3;
+
+  /**
+   * Creates a model for video generation.
+   */
+  video(modelId: GoogleGenerativeAIVideoModelId): Experimental_VideoModelV3;
 
   /**
    * @deprecated Use `chat()` instead.
@@ -110,31 +118,31 @@ Creates a model for image generation.
 
 export interface GoogleGenerativeAIProviderSettings {
   /**
-Use a different URL prefix for API calls, e.g. to use proxy servers.
-The default prefix is `https://generativelanguage.googleapis.com/v1beta`.
+   * Use a different URL prefix for API calls, e.g. to use proxy servers.
+   * The default prefix is `https://generativelanguage.googleapis.com/v1beta`.
    */
   baseURL?: string;
 
   /**
-API key that is being send using the `x-goog-api-key` header.
-It defaults to the `GOOGLE_GENERATIVE_AI_API_KEY` environment variable.
+   * API key that is being send using the `x-goog-api-key` header.
+   * It defaults to the `GOOGLE_GENERATIVE_AI_API_KEY` environment variable.
    */
   apiKey?: string;
 
   /**
-Custom headers to include in the requests.
-     */
+   * Custom headers to include in the requests.
+   */
   headers?: Record<string, string | undefined>;
 
   /**
-Custom fetch implementation. You can use it as a middleware to intercept requests,
-or to provide a custom fetch implementation for e.g. testing.
-    */
+   * Custom fetch implementation. You can use it as a middleware to intercept requests,
+   * or to provide a custom fetch implementation for e.g. testing.
+   */
   fetch?: FetchFunction;
 
   /**
-Optional function to generate a unique ID for each request.
-     */
+   * Optional function to generate a unique ID for each request.
+   */
   generateId?: () => string;
 
   /**
@@ -145,7 +153,7 @@ Optional function to generate a unique ID for each request.
 }
 
 /**
-Create a Google Generative AI provider instance.
+ * Create a Google Generative AI provider instance.
  */
 export function createGoogleGenerativeAI(
   options: GoogleGenerativeAIProviderSettings = {},
@@ -209,6 +217,15 @@ export function createGoogleGenerativeAI(
       fetch: options.fetch,
     });
 
+  const createVideoModel = (modelId: GoogleGenerativeAIVideoModelId) =>
+    new GoogleGenerativeAIVideoModel(modelId, {
+      provider: providerName,
+      baseURL,
+      headers: getHeaders,
+      fetch: options.fetch,
+      generateId: options.generateId ?? generateId,
+    });
+
   const provider = function (modelId: GoogleGenerativeAIModelId) {
     if (new.target) {
       throw new Error(
@@ -235,6 +252,7 @@ export function createGoogleGenerativeAI(
   provider.textEmbeddingModel = createEmbeddingModel;
   provider.image = createImageModel;
   provider.imageModel = createImageModel;
+  provider.video = createVideoModel;
   provider.tools = googleTools;
   provider.files = filesClient;
 
@@ -242,6 +260,6 @@ export function createGoogleGenerativeAI(
 }
 
 /**
-Default Google Generative AI provider instance.
+ * Default Google Generative AI provider instance.
  */
 export const google = createGoogleGenerativeAI();