+++ packages/provider-utils/src/types/model-message.ts
@@ -4,8 +4,8 @@ import { ToolModelMessage } from './tool-model-message';
 import { UserModelMessage } from './user-model-message';
 
 /**
-A message that can be used in the `messages` field of a prompt.
-It can be a user message, an assistant message, or a tool message.
+ * A message that can be used in the `messages` field of a prompt.
+ * It can be a user message, an assistant message, or a tool message.
  */
 export type ModelMessage =
   | SystemModelMessage